+++ src/api/services/query/getDetails.ts
